# frontend/engineer-app/VISTA-Vision-Integrated-Smart-Track-Analytics-aizen/VISTA-Vision-Integrated-Smart-Track-Analytics-aizen/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RailTrack-AI Predictive Maintenance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /components:
    get:
      operationId: listComponents
      tags: [components]
      summary: List track components with optional filters
      parameters:
        - name: hubId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: sectorId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of components
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackComponent"
    post:
      operationId: createComponent
      tags: [components]
      summary: Create a track component
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComponentBody"
      responses:
        "201":
          description: Created component
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackComponent"
...
components:
  schemas:
...
    TrackComponent:
      type: object
      properties:
        id:
          type: integer
        componentId:
          type: string
          description: Human-readable ID like TRK-BNG-MJ-001
        type:
          type: string
          description: "ERC Clip | Rail Joint | Fish Plate | Sleeper | Ballast | Signal Cable"
        hubId:
          type: integer
        sectorId:
          type: integer
        ageMonths:
          type: integer
        lastInspection:
          type: ["string", "null"]
          format: date-time
        ciiScore:
          type: number
        ciiStatus:
          type: string
          description: "NOMINAL | MODERATE | HIGH_RISK | CRITICAL"
        loadStress:
          type: number
          description: Average tonnage (GMT)
        rainfallIndex:
          type: number
          description: Corrosion factor 0-1
        thermalGradient:
          type: number
          description: Heat cycles
        latitude:
          type: number
        longitude:
          type: number
        location:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - componentId
        - type
        - hubId
        - sectorId
        - ageMonths
        - ciiScore
        - ciiStatus
        - loadStress
        - rainfallIndex
        - thermalGradient
        - latitude
        - longitude
        - location
        - createdAt
    CreateComponentBody:
      type: object
      properties:
        componentId:
          type: string
        type:
          type: string
        hubId:
          type: integer
        sectorId:
          type: integer
        ageMonths:
          type: integer
        loadStress:
          type: number
        rainfallIndex:
          type: number
        thermalGradient:
          type: number
        latitude:
          type: number
        longitude:
          type: number
        location:
          type: string
      required:
        - componentId
        - type
        - hubId
        - sectorId
        - ageMonths
        - loadStress
        - rainfallIndex
        - thermalGradient
        - latitude
        - longitude
        - location
...